backend/requirements.txt
Flask==3.1.0
Flask-SQLAlchemy==3.1.1
Flask-JWT-Extended==4.7.1
Flask-CORS==5.0.1
Flask-Limiter==3.12
Flask-Migrate==4.1.0
SQLAlchemy==2.0.36
alembic==1.14.1
psycopg2-binary==2.9.10
pydantic==2.10.6
bcrypt==4.2.1
requests==2.32.3
redis==5.2.1
gunicorn==23.0.0
python-dotenv==1.0.1
pytest==8.3.4
pytest-flask==1.3.0
